+++ os_windows.go
@@ -3,11 +3,11 @@ package main
 import (
 	"cmp"
 	"fmt"
-	"log"
 	"os"
 	"os/exec"
 	"os/user"
 	"path/filepath"
+	"slices"
 	"strings"
 	"syscall"
 
@@ -21,7 +21,7 @@ var (
 	envShell  = os.Getenv("SHELL")
 )
 
-var envPathExt = os.Getenv("PATHEXT")
+var envPathExts []string
 
 var (
 	gDefaultShell       = "cmd"
@@ -115,6 +115,14 @@ func init() {
 		gDefaultSocketPath = filepath.Join(runtime, fmt.Sprintf("lf.%s.sock", gUser.Username))
 		syscall.Close(socket)
 	}
+
+	s := cmp.Or(os.Getenv("PATHEXT"), ".COM;.EXE;.BAT;.CMD")
+	for ext := range strings.SplitSeq(s, ";") {
+		if ext == "" {
+			continue
+		}
+		envPathExts = append(envPathExts, strings.ToLower(ext))
+	}
 }
 
 func detachedCommand(name string, arg ...string) *exec.Cmd {
@@ -174,13 +182,11 @@ func setDefaults() {
 func setUserUmask() {}
 
 func isExecutable(f os.FileInfo) bool {
-	for e := range strings.SplitSeq(envPathExt, string(filepath.ListSeparator)) {
-		if strings.HasSuffix(strings.ToLower(f.Name()), strings.ToLower(e)) {
-			log.Print(f.Name(), e)
-			return true
-		}
+	ext := filepath.Ext(f.Name())
+	if ext == "" {
+		return false
 	}
-	return false
+	return slices.Contains(envPathExts, strings.ToLower(ext))
 }
 
 func isHidden(f os.FileInfo, path string, hiddenfiles []string) bool {